Move all predictions to Sample class
