+++ season.py
@@ -1,15 +1,11 @@
 from typing import Dict
 
 from game import Game
-from prediction import Prediction, Classifier, NeuralNetworkClassifier, FiftyFiftyClassifier, SeedsBasedClassifier
+from classifier import Classifier, SeedsBasedClassifier
 from sample import Sample
 from teams import Team
 from tournament import Tournament
 from seed import Seed
-from statistics import mean
-
-from sklearn.neural_network import MLPClassifier
-from sklearn.preprocessing import MinMaxScaler
 
 
 class Season:
@@ -111,59 +107,37 @@ class Season:
     
     Because a match-up can be viewed as "team_1 vs. team_2" or "team_2 vs. team_1", and the predictions for these
     two point of views always sum up to 1 (no draws), we only output predictions for "team_1 vs. team_2" views, where
-    the team_1's ID is strictly smaller than team_2's ID. That rule is enforced at the Prediction class level.
+    the team_1's ID is strictly smaller than team_2's ID. That rule is enforced at the Sample class level.
     """
-    def predict(self, classifier: Classifier) -> [Prediction]:
-        predictions: [Prediction] = []
-
+    def predict(self, classifier: Classifier) -> [Sample]:
         # Sorted array (ascending) of IDs of teams which participate to this season's NCAA tournament.
         tournament_teams_ids: [int] = self.tournament.team_ids
 
         """
         Go through the upper triangular matrix (without the diagonal: teams don't play themselves!) and predict the
-        winning probability of team_1 vs. team_2. 
-        
-        If n is the number of tournament teams (may change each year, depending on the number of play-in games), 
-        we generate n * (n - 1) / 2 predictions.
+        winning probability of team_1 vs. team_2. Number of match-ups: n * (n - 1) / 2, where n = number of teams. 
         """
+
+        samples: [Sample] = []
         for idx_1, team_1_id in enumerate(tournament_teams_ids):
             # No need to enumerate for second team, we only use that team's ID.
             for team_2_id in tournament_teams_ids[idx_1+1:]:
-                expected_outcome = self.tournament.get_expected_outcome(team_1_id, team_2_id)
-                prediction = self.predict_match_up(classifier, team_1_id, team_2_id, expected_outcome)
-                predictions.append(prediction)
+                samples.append(self.get_sample(team_1_id, team_2_id))
 
-        return predictions
+        classes_probabilities = classifier.predict_proba(samples)
+        for idx, sample in enumerate(samples):
+            sample.win_p = classes_probabilities[idx][1]
+
+        return samples
 
     """
-    Predicts a single match-up winning probability for the team with the smallest ID (team_1). 
-    
-    The argument team_1_id needs to be smaller than team_2_id.
-    
-    The method takes care of running predictions for both visions of the match-up, to validate the approach. 
-    But it always outputs the winning probability from the vision team_1 vs. team_2. 
-    
-    This method is not optimal as it performs a prediction for a single match-up, when we would want to predict for
-    a vector of match-ups simultaneously. If we do get confirmation that the first vision (team_1 vs. team_2) winning
-    probability is equal to the second vision (team_2 vs. team_1) losing probability (algorithm symmetry), then we'll
-    cut the number of calls to predict_proba by 2. Then we will bring the number of calls to predict_proba to 1
-    per season.
+    Get sample for team_1 vs. team_2 match-up.
     """
-    def predict_match_up(self, classifier: Classifier, team_1_id: int, team_2_id: int, expected_outcome: int) -> Prediction:
-        assert team_1_id != team_2_id, f"Cannot predict outcome of {team_1_id} vs. {team_2_id}, they're the " \
-                                       f"same team"
+    def get_sample(self, team_1_id: int, team_2_id: int):
+        expected_outcome = self.tournament.get_expected_outcome(team_1_id, team_2_id)
         match_up_features = self.get_match_up_features(team_1_id, team_2_id)
-        match_up_samples = [Sample(team_1_id, team_2_id, match_up_features[0])]
-        predictions = classifier.predict_proba(match_up_samples)
 
-        # First view is when we consider the team_1 vs. team_2 features: we retrieve the winning probability of team_1.
-        first_view_win_prediction = predictions[0][1]
-
-        # Second view is when we consider the team_2 vs. team_1 features: we retrieve the losing probability of team_2.
-        # Ideally, the two predictions exactly match.
-        # second_view_loss_prediction = predictions[1][0]
-
-        return Prediction(self.year, team_1_id, team_2_id, first_view_win_prediction, expected_outcome)
+        return Sample(team_1_id, team_2_id, match_up_features[0], expected_outcome)
 
     @property
     def year(self):
@@ -245,124 +219,3 @@ class RegularSeason:
         assert won_games + lost_games != 0, f"No game records for team {team_id} during the {self.year-1}-{self.year}" \
                                             f" regular season."
         return won_games / (won_games + lost_games)
-
-
-class Span:
-    """
-    A span wraps a range of seasons, usually consecutive.
-    Spans allow for manipulating several years of game data.
-
-    To most common use of spans is to create two Span instances:
-    - the first to learn the links between regular season and tournament match-ups;
-    - the second to score the validity of those "learned" links.
-
-    E.g. One span from 1985 to 2015 (included) serves as training data.
-         A second span from 2016 to 2021 (included, but without 2020) serves as a
-         validation data set.
-    """
-
-    def __init__(self, seasons: [Season]):
-        self.seasons: [Season] = seasons
-
-    """
-    The train API lets users fit a Multi-Layer Perceptron classifier (using a logarithmic
-    loss function) based on the outcomes of tournament games of each season in the span. 
-    
-    Each season brings its share of match-ups features based on the tournament games that took
-    place that season: the outcome of those games is the label to learn and that we want to
-    predict accurately in the end. It's 1 (resp. 0) if the first (resp. second) team wins. 
-    
-    Careful, the order of teams features matters.
-     
-    By concatenating all the seasons' features together (and similarly for all the labels), we
-    can fit a classifier and return it for prediction purposes.
-    """
-    def train(self, max_iter: int = 1000) -> Classifier:
-        span_features, span_labels = [], []
-
-        # Concatenate each season's features and labels.
-        for season in self.seasons:
-            season_features, season_labels = season.get_season_features_and_labels()
-            span_features.extend(season_features)
-            span_labels.extend(season_labels)
-
-        # All layers with the same size
-        layer_size = len(span_features[0])
-        print(len(span_features))
-        print(len(span_features[0]))
-
-        # Two layers for now.
-        hidden_layer_sizes = (layer_size, layer_size * 2, layer_size * 3, layer_size * 2, layer_size)
-
-        mlp_classifier = MLPClassifier(hidden_layer_sizes=hidden_layer_sizes, max_iter=max_iter)
-
-        # TODO: Scale features for faster convergence.
-        # TODO: Monitor the learning part with loading bar for Jupyter notebook.
-        mlp_classifier.fit(span_features, span_labels)
-
-        return NeuralNetworkClassifier(mlp_classifier)
-
-    """
-    The predict API relies on a dictionary of classifiers indexed by the season's year (a classifier per season)
-    to give predictions. If a specific season doesn't have an entry in that dictionary, we use the 50/50 classifier. 
-    
-    Be careful to always provide a classifier which was fitted on a span which does
-    not overlap with the span you predict on. Otherwise, your results may be biased towards
-    your training seasons.
-    """
-    def predict(self, classifiers: Dict = {}) -> Dict:
-        # Map of season's year to predictions.
-        span_predictions: Dict = {}
-
-        for season in self.seasons:
-            season_classifier = classifiers.get(season.year, FiftyFiftyClassifier())
-            season_predictions = season.predict(season_classifier)
-            span_predictions[season.year] = season_predictions
-
-        return span_predictions
-
-    """
-    Returns exactly each season's seeds based classifier.
-    """
-    def get_seasons_seeds_based_classifiers(self) -> Dict[int, SeedsBasedClassifier]:
-        seasons_seeds_based_classifiers: Dict[int, SeedsBasedClassifier] = {}
-
-        for season in self.seasons:
-            seasons_seeds_based_classifiers[season.year] = season.get_seeds_based_classifier()
-
-        return seasons_seeds_based_classifiers
-
-    """
-    Returns a dictionary mapping each season to the same classifier, the one provided as input.
-    """
-    def build_seasons_classifiers_map(self, classifier: Classifier) -> Dict[int, Classifier]:
-        seasons_classifiers: Dict[int, Classifier] = {}
-
-        for season in self.seasons:
-            seasons_classifiers[season.year] = classifier
-
-        return seasons_classifiers
-
-    """
-    The score API takes in predictions for each of the seasons we want to score.
-    
-    It logs and returns each season's score and an average score for the span. 
-    The output is a dictionary with the season's years as keys (-1 for the special case of the span score).
-    """
-    @staticmethod
-    def score(span_predictions: Dict) -> Dict[int, float]:
-        scores: Dict[int, float] = {}
-
-        for year, season_predictions in span_predictions.items():
-            relevant_predictions = []
-            for prediction in season_predictions:
-                if prediction.expected_outcome != -1:
-                    relevant_predictions.append(prediction)
-
-            relevant_predictions_scores = [prediction.score for prediction in relevant_predictions]
-            scores[year] = mean(relevant_predictions_scores)
-
-        span_score = mean(scores.values())
-        scores[-1] = span_score
-
-        return scores
